Add failing test: local variable as operator
Something to ponder. Right now there's a code path that checks
"is the operator a symbol? then handle the built-in symbolic
things!". Maybe should start by adding a "look up in environment"
case in that function. But what we should really do is move all
the existing built-in operators into the environment.