Write an S-expression parser
Well, that saved an unexpected number of lines of code. What have
we learned here? Don't get too used to writing things the long
and stupid way.

The parser is essentially a CPS-transformed version of the
corresponding recursive parser. I like how it turned out; the only
difficult part was saving the "should we quote?" state from the
left end of a list to the right end (where we finally have the
list and can do the quote wrapping).