Shove primitives into a standard environment
This change felt good; it means that we don't recognize the
built-in primitives by their symbolic name anymore -- instead,
we recognize them by their unique _value_, of a new type
ValuePrimitive.

What's amazing (and not really anticipated) is that this works
even for the non-functional "special forms" `cond`, `label`,
`lambda`, and `quote`. That's because, at the time of KontApp1,
we've evaluated the operator and nothing else. These special
forms take over, and decide which operands to evaluate, if any.
I haven't seen it done this way in a Lisp besides Kernel.