Invent built-in functions
Built-in functions are function-like primitives that bottom out
in a TypeScript body/implementation. They are pure, with the
slight exception that they might throw an exception on a type
mismatch. (`car` and `cdr` do this.)

This takes the load off a number of Kont types, whose sole job
was to facilitate argument evaluation of these primitives, even
though the argument evaluation is completely standard. The
remaining Kont types are much more clearly focused on special
forms.