Allow ValueFunction to have varargs
This way, we can implement 'list' (trivially) as one; it just
returns the argument list unchanged. This is safe, because that
argument list was created just for the function call, and so
is unique.

Also took the chance to remove the corresponding functionality
for ValueBuiltinFunction, as that's now dead weight.